# Conflicts:
#	virtual_machine/src/carriage/operation/_operation_aff.c


Former-commit-id: acecf70e0fb4aeadbdad68437dedfc8d488715f7

+++ virtual_machine/src/vm_params/vmp_player/file/_file.h
@@ -22,8 +22,8 @@
 typedef struct	s_file
 {
 	void		*data;
-	ssize_t		total;
-	ssize_t		capacity;
+	unsigned int		total;
+    unsigned int		capacity;
 	char		*file_name;
 	int			fd;
 	int			is_opened;
